+++ changelog.md
@@ -22,6 +22,9 @@ We're skipping a few version numbers here to keep in sync with the rest of [Micr
     -> send https://github.com/howdyai/botkit/pull/1720
 * FIX: improve mapping of bot framework activity fields
 * quick replies and suggestedActions now procssed as mustache templates
+-> https://github.com/howdyai/botkit/pull/1731
+* keyname for convo.ask and convo.addQuestion can now be set to null to discard answer.
+-> https://github.com/howdyai/botkit/pull/1716
 
 * FIX: issue where child dialog results causes match to bork
     --??